+++ src/UnbreakableBox.cpp
@@ -11,11 +11,11 @@ UnbreakableBox::UnbreakableBox(Drop *drop) : Box(drop) {
 	// 	std::cout << "UnbreakableBox was passed a NULL drop, Constructed\n";
 }
 
-UnbreakableBox::UnbreakableBox(std::pair<int, int> xy) : Box(xy) {
+UnbreakableBox::UnbreakableBox(std::pair<float, float> xy) : Box(xy) {
 	std::cout << "UnbreakableBox at x: " << xy.first << " y: " << xy.second << " Constructed\n";
 }
 
-UnbreakableBox::UnbreakableBox(std::pair<int, int> xy, Drop *drop) : Box(xy, drop) {
+UnbreakableBox::UnbreakableBox(std::pair<float, float> xy, Drop *drop) : Box(xy, drop) {
 	// if (drop)
 	// 	std::cout << "UnbreakableBox with drop at x: " << xy.first << " y: " << xy.second << " Constructed\n";
 	// else
